refactor(exit-ticket): extract local fallback algorithm + data tables (Related to #1908)
- Extract COMMON_PARTICLES, CONFUSABLE_CHARS, LocalQuestion, generateLocalQuestions
  into exitTicketLocalFallback.ts (pure utility, no React deps)
- ExitTicket.tsx reduced from 520L → 308L, imports from new module
- 12 vitest tests: particles skip, confusable distractor, cap-at-3, tautology checks
- Zero TS errors in changed files; vite build passes
